requirements.txt
Flask==0.12
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==1.0
Pillow==4.2.0
pyparsing==2.1.4
python-dateutil==2.5.3
pytz==2016.4
six==1.10.0
Werkzeug==0.12.1